% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{Plotter
Plots covid seasonality by locales}
\usage{
plotter(
  locales = NULL,
  decomposition.method = "twitter",
  anomalize.method = "gesd"
)
}
\arguments{
\item{locales}{the country of interest}

\item{decomposition.method}{the decomposition method for anomalize}

\item{anomalize.method}{the anomalize method for anomalize}
}
\description{
Plotter
Plots covid seasonality by locales
}
